requirements.txt
streamlit
pandas
gTTS
torch
transformers
pillow
